Add files via upload
Roughly edited (yeah, it's pretty bad... it occurs to me now that I didn't cross-fade any of the cuts...) little song demonstrating my new synth module, the Frenetic Robot Noise. The FRN comes in at about 1:30, and I mess with the knobs for most of the rest of the track.